--- src/TestUtils/__init__.py
@@ -1,1 +1,0 @@
-from .LogEngine import *
